E2E: update Vue, add Svelte, cleanup (#470)
* Move playwright-install to the toplevel

This avoids the confusing use of one of the E2E harnesses.

* Document how to run one Playwright E2E manually.

* Reduce duplication in Playwright configs

* Move jsx-types to global.d.ts

including project template

* Vue E2E comsume source code

Earlier Vue testing somewhat worked - but the Vue output was incomplete. For now, the working way to consume the Vue output from Mitosis is as source code, copied in to the test harness for building.

* E2E - add Svelte

This required removing an "inner function", which some other output targets support.